remove: 패키지 이동 및 public 메서드로 설정
